+++ src/main/java/roomescape/controller/exception/GlobalExceptionHandler.java
@@ -1,7 +1,6 @@
 package roomescape.controller.exception;
 
 import jakarta.validation.ConstraintViolationException;
-import java.util.NoSuchElementException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -17,10 +16,12 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-import roomescape.payment.exception.PaymentClientException;
-import roomescape.payment.exception.PaymentServerException;
-import roomescape.security.exception.AccessDeniedException;
-import roomescape.security.exception.UnauthorizedException;
+import roomescape.exception.AccessDeniedException;
+import roomescape.exception.PaymentClientException;
+import roomescape.exception.PaymentServerException;
+import roomescape.exception.UnauthorizedException;
+
+import java.util.NoSuchElementException;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
