+++ src/main/java/roomescape/infrastructure/JwtTokenProvider.java
@@ -6,13 +6,14 @@ import io.jsonwebtoken.Jws;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.UnsupportedJwtException;
 import io.jsonwebtoken.security.Keys;
-import java.nio.charset.StandardCharsets;
-import java.util.Date;
-import javax.crypto.SecretKey;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import roomescape.security.exception.UnauthorizedException;
-import roomescape.security.provider.TokenProvider;
+import roomescape.exception.UnauthorizedException;
+import roomescape.service.TokenProvider;
+
+import javax.crypto.SecretKey;
+import java.nio.charset.StandardCharsets;
+import java.util.Date;
 
 @Component
 public class JwtTokenProvider implements TokenProvider {
