+++ src/main/java/roomescape/service/TokenProvider.java
@@ -1,4 +1,4 @@
-package roomescape.security.provider;
+package roomescape.service;
 
 public interface TokenProvider {
