+++ src/test/java/roomescape/service/AuthServiceTest.java
@@ -1,20 +1,19 @@
 package roomescape.service;
 
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.doReturn;
-
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.SpyBean;
 import roomescape.security.authentication.Authentication;
-import roomescape.security.provider.TokenProvider;
 import roomescape.service.dto.request.CreateMemberRequest;
 import roomescape.service.dto.request.CreateTokenRequest;
 import roomescape.service.dto.response.TokenResponse;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.doReturn;
 
 class AuthServiceTest extends BaseServiceTest {
