+++ src/test/java/roomescape/support/FakePaymentClient.java
@@ -1,6 +1,6 @@
 package roomescape.support;
 
-import roomescape.payment.PaymentClient;
+import roomescape.service.PaymentClient;
 import roomescape.service.dto.request.PaymentRequest;
 
 public class FakePaymentClient implements PaymentClient {
